pyproject.toml
[project]
name = "week3-llamaindex-crewai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai[tools]>=0.157.0",
    "google-generativeai>=0.8.5",
    "llama-index>=0.13.1",
    "llama-index-llms-google-genai>=0.3.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]
